feat: P2 fixes - Uint64 type and middleware tests
P2 Fixes:
- logger.go: Added Uint64Type and Uint64() constructor for large unsigned values
- zap.go: Added Uint64Type case to convertField
- fields/blockchain.go: Updated BlockHeight, Slot, Epoch, Nonce, GasLimit, GasPrice, GasUsed to use Uint64
- middleware/http/http_test.go: Added tests for Handler, Client, Transport, WithFilter
- middleware/grpc/grpc_test.go: Added tests for ServerHandler, ClientHandler, WithFilter

This prevents truncation of uint64 values > 2^63 in blockchain fields.

+++ fields/blockchain.go
@@ -42,7 +42,7 @@ func TxType(txType string) ion.Field {
 
 // BlockHeight creates a block height field.
 func BlockHeight(height uint64) ion.Field {
-	return ion.Int64("block_height", int64(height))
+	return ion.Uint64("block_height", height)
 }
 
 // BlockHash creates a block hash field.
@@ -52,12 +52,12 @@ func BlockHash(hash string) ion.Field {
 
 // Slot creates a slot number field (Solana).
 func Slot(slot uint64) ion.Field {
-	return ion.Int64("slot", int64(slot))
+	return ion.Uint64("slot", slot)
 }
 
 // Epoch creates an epoch field.
 func Epoch(epoch uint64) ion.Field {
-	return ion.Int64("epoch", int64(epoch))
+	return ion.Uint64("epoch", epoch)
 }
 
 // --- Chain Context Fields ---
@@ -210,22 +210,22 @@ func ErrorMsg(err error) ion.Field {
 
 // Nonce creates a transaction nonce field.
 func Nonce(n uint64) ion.Field {
-	return ion.Int64("nonce", int64(n))
+	return ion.Uint64("nonce", n)
 }
 
 // GasLimit creates a gas limit field.
 func GasLimit(limit uint64) ion.Field {
-	return ion.Int64("gas_limit", int64(limit))
+	return ion.Uint64("gas_limit", limit)
 }
 
 // GasPrice creates a gas price field (in smallest unit, e.g., wei).
 func GasPrice(price uint64) ion.Field {
-	return ion.Int64("gas_price", int64(price))
+	return ion.Uint64("gas_price", price)
 }
 
 // GasUsed creates a gas used field.
 func GasUsed(used uint64) ion.Field {
-	return ion.Int64("gas_used", int64(used))
+	return ion.Uint64("gas_used", used)
 }
 
 // Value creates a transaction value field.
